requirements.txt
# This file is autogenerated. Do not edit it manually.
# BEGIN: `requirements-core.txt`
# Google Cloud Function pre-installed packages: https://cloud.google.com/functions/docs/concepts/python-runtime

click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pip==18.0
# requests==2.19.1
setuptools==40.2.0
Werkzeug==0.14.1
wheel==0.31.1

# Custom/Override packages
backoff==1.8.0
python-dateutil==2.7.5
requests==2.21.0
# END: `requirements-core.txt`



## The following requirements were added by pip freeze:
certifi==2018.11.29
chardet==3.0.4
idna==2.8
six==1.12.0
urllib3==1.24.1
